+++ apps/api/prisma/migrations/20240107110836_add_line_number_start/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "SnippetEditorTab" ADD COLUMN     "lineNumberStart" INTEGER NOT NULL DEFAULT 1;