+++ apps/api/src/modules/project/mapper/get-project-by-id-mapper.ts
@@ -46,6 +46,7 @@ export function createCompleteProjectGetByIdResponseMapper(
       languageId: editor.languageId,
       code: editor.code,
       tabName: editor.tabName,
+      lineNumberStart: editor.lineNumberStart,
     })),
     isOwner: false,
   };