+++ apps/api/src/modules/project/schema/project-update.schema.ts
@@ -22,6 +22,7 @@ export const SnippetEditorTabsUpdateRequestSchema = Type.Array(
       code: Type.String(),
       languageId: Type.String(),
       tabName: Type.String(),
+      lineNumberStart: Type.Integer({minimum: 1, maximum: 999_999}),
     },
     {title: 'SnippetEditorTabUpdateRequest'},
   ),