+++ apps/api/src/modules/project/schema/project.schema.ts
@@ -22,6 +22,7 @@ export const BaseSnippetEditorTabsSchema = Type.Array(
     code: Type.String(),
     languageId: Type.String(),
     tabName: Type.String(),
+    lineNumberStart: Type.Integer({minimum: 1, maximum: 999_999}),
   }),
 );