+++ apps/codeimage/src/core/constants/umami.ts
@@ -7,22 +7,22 @@ declare global {
 const isDev = import.meta.env.DEV;
 
 function getUmamiMock() {
-  const umamiMock: Umami = () => void 0;
+  const umamiMock: Umami = {} as Umami;
 
   if (isDev) {
-    umamiMock.trackEvent = (event_value, event_type, url, website_id) => {
+    umamiMock.track = ((
+      event_name: string,
+      event_data?: {[key: string]: string | number},
+    ) => {
       console.groupCollapsed(`[DEV] Umami track event`);
-      console.table([{event_value, event_type, website_id, url}]);
+      console.log({
+        event_name,
+        event_data,
+      });
       console.groupEnd();
-    };
-    umamiMock.trackView = (url, referrer, website_id) => {
-      console.groupCollapsed(`[DEV] Umami track view`);
-      console.table([{url, referrer, website_id}]);
-      console.groupEnd();
-    };
+    }) as typeof umamiMock.track;
   } else {
-    umamiMock.trackEvent = () => void 0;
-    umamiMock.trackView = () => void 0;
+    umamiMock.track = () => void 0;
   }
 
   return umamiMock;