+++ apps/codeimage/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @codeimage/app
 
+## 1.5.1
+
+### Patch Changes
+
+- [#620](https://github.com/riccardoperra/codeimage/pull/620) [`8046d566`](https://github.com/riccardoperra/codeimage/commit/8046d566b2e9542a8d0c6d66c156afd396713746) Thanks [@riccardoperra](https://github.com/riccardoperra)! - feat: update umami
+
+- Updated dependencies [[`1f9e5788`](https://github.com/riccardoperra/codeimage/commit/1f9e5788e1a011542f850d2abd35f12201a19b8b)]:
+  - @codeimage/config@8.2.5
+
 ## 1.5.0
 
 ### Minor Changes